requirements.txt
azure-common==1.1.26
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
asgiref==3.3.1
bleach==3.3.0
dj-database-url==0.5.0
Django==3.1.6
django-dotenv==1.4.2
django-markdownify==0.8.2
django-markdownx==3.0.1
django-widget-tweaks==1.4.8
django-storages==1.11.1
gunicorn==20.0.4
Markdown==3.3.3
packaging==20.9
Pillow==8.1.0
psycopg2==2.8.6
pyparsing==2.4.7
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
webencodings==0.5.1
whitenoise==5.2.0
